+++ frontend/src/App.js
@@ -6,6 +6,7 @@ import Admin from "./pages/Admin";
 import SponsorStatus from "./pages/SponsorStatus";
 import SponsorChains from "./pages/SponsorChains";
 import SponsorKit from "./pages/SponsorKit";
+import PressKit from "./pages/PressKit";
 import RarePreview from "./pages/RarePreview";
 import Rituals from "./pages/Rituals";
 import Journal from "./pages/Journal";
@@ -55,6 +56,7 @@ function App() {
           <Route path="/sponsor/cancelled" element={<SponsorStatus cancelled />} />
           <Route path="/sponsor/chains" element={<SponsorChains />} />
           <Route path="/sponsors" element={<SponsorKit />} />
+          <Route path="/press" element={<PressKit />} />
           <Route path="/dev/rare" element={<RarePreview />} />
           <Route path="/terms" element={<LegalPage type="terms" />} />
           <Route path="/privacy" element={<LegalPage type="privacy" />} />